fix: change Service.UptimeSecs property from float64 to int64

+++ src/domain/entity/service.go
@@ -6,7 +6,7 @@ type Service struct {
 	Name            valueObject.ServiceName   `json:"name"`
 	Status          valueObject.ServiceStatus `json:"status"`
 	Pids            *[]uint32                 `json:"pids,omitempty"`
-	UptimeSecs      *float64                  `json:"uptimeSecs,omitempty"`
+	UptimeSecs      *int64                    `json:"uptimeSecs,omitempty"`
 	CpuUsagePercent *float64                  `json:"cpuUsagePercent,omitempty"`
 	MemUsagePercent *float32                  `json:"memUsagePercent,omitempty"`
 }
@@ -15,7 +15,7 @@ func NewService(
 	name valueObject.ServiceName,
 	status valueObject.ServiceStatus,
 	pids *[]uint32,
-	uptimeSecs *float64,
+	uptimeSecs *int64,
 	cpuUsagePercent *float64,
 	memUsagePercent *float32,
 ) Service {
